requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
Django==6.0.3
djangorestframework==3.17.1
fonttools==4.62.1
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.33.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
